% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonifier.R
\name{buttonifier}
\alias{buttonifier}
\title{Create a button for gene symbols in an Rmd}
\usage{
buttonifier(
  df,
  new_cols = c("GC", "UNIPROT"),
  col_to_use = "SYMBOL",
  output_format = "DT"
)
}
\arguments{
\item{df}{A dataframe with at least on column with gene Symbols named: SYMBOL}

\item{new_cols}{At least one of: "GC", "NCBI", "GTEX", "UNIPROT", "dbPTM", "HPA" "PUBMED"}

\item{col_to_use}{name of the columns were the gene symbols are stored. Default is SYMBOL}

\item{output_format}{a parameter deciding which output format to return, either a DT:datatable (recommended)
or a simple dataframe (DF).In the latter case it is important that if the data is visualized with the
\code{datatable} function the parameter escape must be set to FALSE}
}
\value{
A dataframe or a \code{DT} datatable object with columns adding HTML objects that link to websites with further information on the genes in question.
}
\description{
A function to turn Gene Symbols into buttons in an Rmd linking to various Portals for further info
about these genes.
Current supported portals are: GeneCards, NCBI, GTEx, Uniprot, dbPTM, Human Protein Atlas
}
\examples{
data(res_airway, package = "mosdef")
res_de <- res_airway
res_df <- deseqresult2df(res_de)
# Subsetting for quicker run
res_df <- res_df[1:100,]
buttonifier(res_df)
}
