% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_dbPTM}
\alias{create_link_dbPTM}
\title{Link to dbPTM database}
\usage{
create_link_dbPTM(val)
}
\arguments{
\item{val}{Character, the gene symbol}
}
\value{
HTML for an action button
}
\description{
Link to dbPTM database
}
