% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_ENS}
\alias{create_link_ENS}
\title{Link to Ensemble database}
\usage{
create_link_ENS(val, species = "Mus_musculus")
}
\arguments{
\item{val}{Character, the gene symbol}

\item{species}{The species to be analyzed e.g "Mus_musculus"}
}
\value{
HTML for an action button
}
\description{
Link to Ensemble database
}
