% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_genecards}
\alias{create_link_genecards}
\title{Link to the GeneCards database}
\usage{
create_link_genecards(val)
}
\arguments{
\item{val}{Character, the gene symbol of interest}
}
\value{
HTML for an action button
}
\description{
Link to the GeneCards database
}
