% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_GO}
\alias{create_link_GO}
\title{Link to AMIGO database}
\usage{
create_link_GO(val)
}
\arguments{
\item{val}{Character, the GOID}
}
\value{
HTML for an action button
}
\description{
Link to AMIGO database
}
