% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_HPA}
\alias{create_link_HPA}
\title{Link to the Human Protein Atlas}
\usage{
create_link_HPA(val)
}
\arguments{
\item{val}{Character, the gene symbol}
}
\value{
HTML for an action button
}
\description{
Link to the Human Protein Atlas
}
\examples{
create_link_HPA("Oct4")

}
