% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_pubmed}
\alias{create_link_pubmed}
\title{Link to Pubmed}
\usage{
create_link_pubmed(val)
}
\arguments{
\item{val}{Character, the gene symbol}
}
\value{
HTML for an action button
}
\description{
Link to Pubmed
}
