% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqresult2df.R
\name{deseqresult2df}
\alias{deseqresult2df}
\title{Generate a table from the \code{DESeq2} results}
\usage{
deseqresult2df(res_de, FDR = NULL)
}
\arguments{
\item{res_de}{A \code{DESeqResults} object.}

\item{FDR}{Numeric value, specifying the significance level for thresholding
adjusted p-values. Defaults to NULL, which would return the full set of results
without performing any subsetting based on FDR.}
}
\value{
A tidy \code{data.frame} with the results from differential expression,
sorted by adjusted p-value. If FDR is specified, the table contains only genes
with adjusted p-value smaller than the value.
}
\description{
Generate a tidy table with the results of \code{DESeq2}
}
\examples{
library(airway)
library(DESeq2)
data(airway)
airway
dds_airway <- DESeqDataSet(airway, design= ~ cell + dex)
# Example, performing extraction of enriched functional categories in
# detected significantly expressed genes
dds_airway <- DESeq(dds_airway)
res_airway <- results(dds_airway)
head(res_airway)
res_df <- deseqresult2df(res_airway)
head(res_df)
}
