% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseqresult2tbl.R
\name{deseqresult2tbl}
\alias{deseqresult2tbl}
\title{Generate a tidy table with the results of DESeq}
\usage{
deseqresult2tbl(res_de)
}
\arguments{
\item{res_de}{A \code{\link{DESeqResults}} object}
}
\value{
A "tidy" data.frame with all genes
}
\description{
Generate a tidy table with the results of DESeq
}
\examples{

# with simulated data...
library(DESeq2)
data(res_airway, package = "mosdef")
deseqresult2tbl(res_airway)
}
