% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{.get_expr_values.DESeqDataSet}
\alias{.get_expr_values.DESeqDataSet}
\title{get_expr_values, from a DESeqDataSet object}
\usage{
.get_expr_values.DESeqDataSet(de_container, gene, intgroup, assay, normalized)
}
\description{
get_expr_values, from a DESeqDataSet object
}
\keyword{internal}
