% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{.get_expr_values.DGEList}
\alias{.get_expr_values.DGEList}
\title{get_expr_values, from a DGEList object}
\usage{
.get_expr_values.DGEList()
}
\description{
get_expr_values, from a DGEList object
}
\keyword{internal}
