% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{.get_expr_values.EList}
\alias{.get_expr_values.EList}
\title{get_expr_values, from an Elist object}
\usage{
.get_expr_values.EList()
}
\description{
get_expr_values, from an Elist object
}
\keyword{internal}
