% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_symbol_buttons.R
\name{gene_symbol_buttons}
\alias{gene_symbol_buttons}
\title{Create a button for gene symbols in an Rmd}
\usage{
gene_symbol_buttons(
  df,
  new_cols = c("GC", "UNIPROT"),
  col_to_use = "SYMBOL",
  output_format = "DT"
)
}
\arguments{
\item{df}{A dataframe with at least on column with gene Symbols named: SYMBOL}

\item{new_cols}{At least one of: "GC", "NCBI", "GTEX", "UNIPROT", "dbPTM", "HPA"}

\item{col_to_use}{name of the columns were the gene symbols are stored. Default is SYMBOL}

\item{output_format}{a parameter deciding which output format to return, either a DT:datatable (recommended)
or a simple dataframe (DF).In the latter case it is important that if the data is visualized with the
\code{datatable} function the parameter escape must be set to FALSE}
}
\value{
TODO
}
\description{
A function to turn Gene Symbols into buttons in an Rmd linking to various Portals for further info
about these genes.
Current supported portals are: GeneCards, NCBI, GTEX, Uniprot, dbPTM, Human Protein Atlas
}
\examples{
library(dplyr)
library(DESeq2)
data(res_airway, package = "mosdef")
res_de <- res_airway
res_df <- deseqresult2df(res_de)
# Subsetting for quicker run
res_df <- res_df[1:100,]
gene_symbol_buttons(res_df)
}
