% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdef-pkg.R
\docType{package}
\name{mosdef-pkg}
\alias{mosdef}
\alias{mosdef-package}
\alias{mosdef-pkg}
\title{mosdef: mostly useful differential expression functions}
\description{
mostly useful differential expression functions
}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

Authors:
\itemize{
  \item Leon Dammer \email{dammerle@uni-mainz.de} (\href{https://orcid.org/0009-0008-4132-7639}{ORCID})
}

}
