% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdef_check_inputs.R
\name{mosdef_dds_check}
\alias{mosdef_dds_check}
\title{A function checking if your dds contains everything you need}
\usage{
mosdef_dds_check(dds, verbose = FALSE)
}
\arguments{
\item{dds}{A DESeqDataset object created using \code{DESeq2}}

\item{verbose}{Logical, whether to add messages telling the user which steps were taken}
}
\value{
An invisible \code{NULL} after performing the checks
}
\description{
A function checking if your dds contains everything you need
}
\examples{
library("macrophage")
library("DESeq2")
data(gse, package = "macrophage")

dds_macrophage <- DESeqDataSet(gse, design = ~ line + condition)
rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)
keep <- rowSums(counts(dds_macrophage) >= 10) >= 6
dds_macrophage <- dds_macrophage[keep, ]
dds_macrophage <- DESeq(dds_macrophage)

mosdef_dds_check(dds_macrophage)
}
