% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdef_check_inputs.R
\name{mosdef_res_check}
\alias{mosdef_res_check}
\title{A function checking if your res_de contains everything you need}
\usage{
mosdef_res_check(res_de, verbose = FALSE)
}
\arguments{
\item{res_de}{A DESeqResults object created using \code{DESeq2}}

\item{verbose}{Logical, whether to add messages telling the user which steps were taken}
}
\value{
An invisible \code{NULL} after performing the checks
}
\description{
A function checking if your res_de contains everything you need
}
\examples{
data(res_de_macrophage, package = "mosdef")

mosdef_res_check(res_macrophage_IFNg_vs_naive)
}
