% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdef-data.R
\docType{data}
\name{res_airway}
\alias{res_airway}
\title{A sample \code{DESeqResults} object}
\description{
A sample \code{DESeqResults} object, generated in the \code{DESeq2} framework
}
\details{
This \code{DESeqResults} object is on the data from the \code{airway} package

Details on how this object has been created are included in the \code{create_mosdef_data}
script, included in the \code{scripts} folder of the \code{mosdef} package.

This \code{DESeqDataset} object is on the data from the \code{airway} package

Details on how this object has been created are included in the \code{create_mosdef_data}
script, included in the \code{scripts} folder of the \code{mosdef} package.

This enrichment object is on the data from the \code{airway} package

Details on how this object has been created are included in the \code{create_mosdef_data}
script, included in the \code{scripts} folder of the \code{mosdef} package.
}
\references{
TODO

TODO

TODO
}
