% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdef-data.R
\docType{data}
\name{topgoDE_airway}
\alias{topgoDE_airway}
\title{A sample enrichment object}
\description{
A sample enrichment object, generated in the \code{mosdef} and \code{topGO} framework
}
\details{
This enrichment object is on the data from the \code{airway} package

Details on how this object has been created are included in the \code{create_mosdef_data}
script, included in the \code{scripts} folder of the \code{mosdef} package.
}
\references{
TODO
}
