% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluProtable.R
\name{cluproTable}
\alias{cluproTable}
\title{Title}
\usage{
cluproTable(
  res_de = NULL,
  dds = NULL,
  de_genes = NULL,
  bg_genes = NULL,
  mapping,
  de_type = "up_and_down",
  ...
)
}
\arguments{
\item{res_de}{A DESeqResults object created using \code{DESeq2}}

\item{dds}{A DESeqDataset object created using \code{DESeq2}}

\item{de_genes}{A vector of (differentially expressed) genes}

\item{bg_genes}{A vector of background genes, e.g. all (expressed) genes in the assays}

\item{mapping}{Which \code{org.XX.eg.db} to use for annotation - select according to the species}

\item{de_type}{One of: 'up', 'down', or 'up_and_down' Which genes to use for GOterm calculations:}

\item{...}{Further parameters to use for the go_enrich function from \code{clusterProfiler}}
}
\value{
A table containing the computed GO Terms and related enrichment scores
}
\description{
Title
}
\examples{
library(airway)
library(DESeq2)
data(airway)
airway
dds_airway <- DESeqDataSet(airway, design = ~ cell + dex)
# Example, performing extraction of enriched functional categories in
# detected significantly expressed genes
\dontrun{
dds_airway <- DESeq(dds_airway)
res_airway <- results(dds_airway)
library("AnnotationDbi")
library("org.Hs.eg.db")
library(clusterProfiler)
CluProde_airway <- cluproTable(
  res_de = res_airway,
  dds = dds_airway,
  mapping = "org.Hs.eg.db"
)
}

}
