% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_GeneCards}
\alias{create_link_GeneCards}
\title{Link to the GeneCards database}
\usage{
create_link_GeneCards(val)
}
\arguments{
\item{val}{Character, the gene symbol of interest}
}
\value{
HTML for an action button
}
\description{
Link to the GeneCards database
}
\examples{
 create_link_GeneCards("Oct4")

data(res_de_macrophage, package = "mosdef")
res_macrophage_IFNg_vs_naive$SYMBOL <-
   create_link_GeneCards(res_macrophage_IFNg_vs_naive$SYMBOL)
}
