% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_GTEX}
\alias{create_link_GTEX}
\title{Link to the GTEx Portal}
\usage{
create_link_GTEX(val)
}
\arguments{
\item{val}{Character, the gene symbol of interest}
}
\value{
HTML for an action button
}
\description{
Link to the GTEx Portal
}
\examples{
create_link_GTEX("Oct4")

}
