% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_NCBI}
\alias{create_link_NCBI}
\title{Link to NCBI database}
\usage{
create_link_NCBI(val)
}
\arguments{
\item{val}{Character, the gene symbol}
}
\value{
HTML for an action button
}
\description{
Link to NCBI database
}
