% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{create_link_PubMed}
\alias{create_link_PubMed}
\title{Link to Pubmed}
\usage{
create_link_PubMed(val)
}
\arguments{
\item{val}{Character, the gene symbol}
}
\value{
HTML for an action button
}
\description{
Link to Pubmed
}
\examples{
create_link_PubMed("Oct4")

data(res_de_macrophage, package = "mosdef")
res_macrophage_IFNg_vs_naive$SYMBOL <-
  create_link_PubMed(res_macrophage_IFNg_vs_naive$SYMBOL)
}
