% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_symbol_buttons.R
\name{gene_symbol_buttons}
\alias{gene_symbol_buttons}
\title{Create a button for genesymbols in an RMD}
\usage{
gene_symbol_buttons(df, new_cols = c("GC", "UNIPROT"))
}
\arguments{
\item{df}{A dataframe with at least on coloumn with gene Symbols named: SYMBOL}

\item{new_cols}{At least one of: "GC", "NCBI", "GTEX", "UNIPROT"; "dbPTM", "HPA"}
}
\description{
A function to turn Gene Symbols into buttons in an RMD linking to various Portals for further info
about these genes.
Current supported portals are: Genecards, NCBI, GTEX, Uniprot, dbPTM, Human Protein Atlas
}
\examples{
data("gse", package = "macrophage")

dds_macrophage <- DESeqDataSet(gse, design = ~line + condition)
rownames(dds_macrophage) <- substr(rownames(dds_macrophage), 1, 15)
dds_macrophage
keep <- rowSums(counts(dds_macrophage) >= 10) >= 6
dds_macrophage <- dds_macrophage[keep, ]
dds_macrophage
dds_macrophage <- DESeq(dds_macrophage)

res_macrophage_IFNg_vs_naive <- results(dds_macrophage,
                                        contrast = c("condition", "IFNg", "naive"),
                                        lfcThreshold = 1, alpha = 0.05)
res_macrophage_IFNg_vs_naive$SYMBOL <- rowData(dds_macrophage)$SYMBOL
res_df <- as.data.frame(res_macrophage_IFNg_vs_naive@listData)
res_df <-res_df[1:100,]
res_df <- buttons(res_df)
DT::datatable(res_df)
}
