% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_volcano.R
\name{go_volcano}
\alias{go_volcano}
\title{Generates a volcano plot using ggplot2
This function generates a base volcanoplot highlighting genes associated with a certain GOterm
that can then be expanded upon using further ggplot functions.}
\usage{
go_volcano(
  res_de,
  res_enrich,
  mapping = "org.Hs.eg.db",
  term_index,
  L2FC_cutoff = 1,
  col_to_use = NULL,
  enrich_col = "genes",
  down_col = "black",
  up_col = "black",
  highlight_col = "tomato",
  overlaps = 20
)
}
\arguments{
\item{res_de}{A DESeqResults object created using \code{DESeq2}}

\item{res_enrich}{A enrichment result object created by for example using topGOtable}

\item{mapping}{Which \code{org.XX.eg.db} to use for annotation - select according to the species}

\item{term_index}{The location (row) of your GO term of interest in your enrichment result}

\item{L2FC_cutoff}{A numeric value that sets the cutoff for the xintercept argument of ggplot}

\item{col_to_use}{The column in your differential expression results containing your gene symbols.
If you don't have one it is created automatically}

\item{enrich_col}{column name from your res_enrich where the genes associated with your GOterm are stored (for example see the topGOtable result in mosdef)}

\item{down_col}{The colour for your downregulated genes, default is "gray"}

\item{up_col}{The colour for your upregulated genes, default is "gray"}

\item{highlight_col}{The colour for the genes associated with your GOterm default is "tomato"}

\item{overlaps}{number of overlaps ggrepel is supposed to allow when labelling
(for more info check ggrepel documentation)}
}
\value{
A  \code{ggplot2} volcano plot object that can be extended upon by the user
}
\description{
Generates a volcano plot using ggplot2
This function generates a base volcanoplot highlighting genes associated with a certain GOterm
that can then be expanded upon using further ggplot functions.
}
\examples{
library(ggplot2)
library(RColorBrewer) # for a colourful plot
library(ggrepel) # for nice annotations
library(DESeq2)
library("topGO")
library(AnnotationDbi)
library("org.Hs.eg.db")


data(res_de_macrophage, package = "mosdef")
data(res_enrich_macrophage_topGO, package = "mosdef")
p <- go_volcano(
  res_macrophage_IFNg_vs_naive,
  res_enrich = res_enrich_macrophage_topGO,
  term_index = 1,
  L2FC_cutoff = 1,
  mapping = "org.Hs.eg.db",
  overlaps = 20
)
p
}
