% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdef_check_inputs.R
\name{mosdef_dds_check}
\alias{mosdef_dds_check}
\title{A function checking if your dds contains everything you need}
\usage{
mosdef_dds_check(dds, verbose = FALSE)
}
\arguments{
\item{dds}{A DESeqDataset object created using \code{DESeq2}}

\item{verbose}{Logical, whether to add messages telling the user which steps were taken}
}
\value{
@return An invisible \code{NULL} after performing the checks on the \code{DESeqDataset} object;
if not all conditions are satisfied, this function returns messages stating which elements
are missing
}
\description{
A function checking if your dds contains everything you need
}
\examples{
data(dds_airway, package = "mosdef")
mosdef_dds_check(dds_airway)
}
